add feature "simulator"
