+++ .github/workflows/rust.yml
@@ -15,12 +15,19 @@ jobs:
 
     steps:
     - uses: actions/checkout@v2
-    - name: Build
-      run: cargo build --verbose
-    - name: Run tests
-      run: cargo test --verbose
+    - name: Setup
+      run: |
+        rustup toolchain install 1.51.0-x86_64-unknown-linux-gnu
+        rustup component add rustfmt --toolchain 1.51.0-x86_64-unknown-linux-gnu
+        rustup component add clippy --toolchain 1.51.0-x86_64-unknown-linux-gnu
+    - name: Build (default)
+      run: cargo build --all-targets --verbose
+    - name: Build (all features)
+      run: cargo build --all-features --all-targets --verbose
+    - name: Run tests (all features)
+      run: cargo test --all-features --verbose
     - name: Run LibraBFT simulator
-      run: RUST_LOG=warn cargo run --bin librabft_simulator
+      run: RUST_LOG=warn cargo run --features simulator --bin librabft_simulator
     - name: Rustfmt
       run: cargo fmt -- --check
     - name: Clippy
