+++ librabft-v2/Cargo.toml
@@ -8,6 +8,9 @@ license = "Apache-2.0"
 publish = false
 edition = "2018"
 
+[features]
+simulator = ["bft-lib/simulator"]
+
 [dependencies]
 env_logger = "0.8.1"
 failure = "0.1.5"
@@ -15,10 +18,12 @@ log = "0.4.6"
 rand = "0.7.3"
 clap = "2.33"
 csv = "1.1"
-bft-lib = { path = "../bft-lib" }
 futures = "0.3.14"
+
+bft-lib = { path = "../bft-lib" }
 
 [[bin]]
 name = "librabft_simulator"
+required-features = ["simulator"]
 path = "src/main.rs"
 test = true
