+++ librabft-v2/src/lib.rs
@@ -1,11 +1,20 @@
 // Copyright (c) Calibra Research
 // SPDX-License-Identifier: Apache-2.0
 
+/// Type definitions
 pub mod base_types;
+/// Implementation of bft_lib::DataSyncNode
 pub mod data_sync;
+/// Main node state and implementation of bft_lib::ConsensusNode
 pub mod node;
+/// Liveness module.
 pub mod pacemaker;
+/// Blocks, Votes, Quorum Certificates, etc.
 pub mod record;
+/// In-memory index of records.
 pub mod record_store;
-pub mod simulated_context;
+/// Requirements for the external modules provided by `Context`.
 pub mod smr_context;
+
+#[cfg(feature = "simulator")]
+pub mod simulated_context;
